Create Tokenizer Assignment
LLMs much like any other computer software are incapable of truly understanding human languages. This is because computer hardware must receive information in its native language, numbers and logical states. The most well known example of this is binary, however there are many other types of computer-based language such as barcodes. In the case of LLM’s the manner in which information is stored and evaluated is numerical values called, “tokens”. Using a tokenizer it is possible to take standard text and convert it into a token, which can then be read by an LLM. The capabilities of LLM’s are further enhanced when using the technique, “Mixture of Experts”, which is when data is divided amongst multiple specialized LLM models in order to have a more balanced field of expertise. When possessing balanced capabilities, LLMs can be used for a multitude of tasks including coding assistance. For instance, LLMs can assist in tasks such as fixing syntax errors and writing small code snippets.